Move MessageDeleted to Messages namespace
